requirements.txt
attrs==25.4.0
Automat==25.4.16
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
contourpy==1.3.2
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
defusedxml==0.7.1
filelock==3.20.1
fonttools==4.61.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.0
itemloaders==1.3.2
jmespath==1.0.1
joblib==1.5.3
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
narwhals==2.14.0
nltk==3.9.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parsel==1.10.0
pillow==12.0.0
plotly==6.5.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
queuelib==1.8.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
Scrapy==2.13.4
service-identity==24.2.0
six==1.17.0
tldextract==5.3.0
tomli==2.3.0
tqdm==4.67.1
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
w3lib==2.3.1
wordcloud==1.9.4
zope.interface==8.1.1
